re structure the app
